+++ src/alloc/allocator.rs
@@ -4,6 +4,8 @@ use crate::core::def::{NSTDAny, NSTDErrorCode};
 #[repr(C)]
 #[derive(Clone, Copy, Default)]
 pub struct NSTDAllocator {
+    /// Set to nonzero if an error occurs.
+    pub errc: NSTDErrorCode,
     /// Allocates a new block of memory.
     /// Parameters:
     ///     `NSTDAny this` - A pointer to the owner of the allocator.
@@ -22,16 +24,13 @@ pub struct NSTDAllocator {
     ///     `NSTDAny *ptr` - Pointer to the block of memory.
     ///     `NSTDUSize size` - The current size of the block of memory.
     ///     `NSTDUSize new_size` - The new size of the block of memory.
-    /// Returns: `NSTDErrorCode errc` - Nonzero on error.
-    pub reallocate:
-        Option<unsafe extern "C" fn(NSTDAny, &mut NSTDAny, usize, usize) -> NSTDErrorCode>,
+    pub reallocate: Option<unsafe extern "C" fn(NSTDAny, &mut NSTDAny, usize, usize)>,
     /// Deallocates a block of memory.
     /// Parameters:
     ///     `NSTDAny this` - A pointer to the owner of the allocator.
     ///     `NSTDAny *ptr` - Pointer to the block of memory.
     ///     `NSTDUSize size` - Number of bytes to deallocate.
-    /// Returns: `NSTDErrorCode errc` - Nonzero on error.
-    pub deallocate: Option<unsafe extern "C" fn(NSTDAny, &mut NSTDAny, usize) -> NSTDErrorCode>,
+    pub deallocate: Option<unsafe extern "C" fn(NSTDAny, &mut NSTDAny, usize)>,
 }
 
 /// Returns the default memory allocator.
@@ -40,6 +39,7 @@ pub struct NSTDAllocator {
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_alloc_allocator_default() -> NSTDAllocator {
     NSTDAllocator {
+        errc: 0,
         allocate: Some(allocate),
         allocate_zeroed: Some(allocate_zeroed),
         reallocate: Some(reallocate),
@@ -61,17 +61,14 @@ unsafe extern "C" fn allocate_zeroed(_: NSTDAny, size: usize) -> NSTDAny {
 
 /// Default reallocate function.
 #[inline]
-unsafe extern "C" fn reallocate(
-    _: NSTDAny,
-    ptr: &mut NSTDAny,
-    size: usize,
-    new_size: usize,
-) -> NSTDErrorCode {
-    crate::alloc::nstd_alloc_reallocate(ptr, size, new_size)
+unsafe extern "C" fn reallocate(this: NSTDAny, ptr: &mut NSTDAny, size: usize, new_size: usize) {
+    let this = this as *mut NSTDAllocator;
+    (*this).errc |= crate::alloc::nstd_alloc_reallocate(ptr, size, new_size);
 }
 
 /// Default deallocate function.
 #[inline]
-unsafe extern "C" fn deallocate(_: NSTDAny, ptr: &mut NSTDAny, size: usize) -> NSTDErrorCode {
-    crate::alloc::nstd_alloc_deallocate(ptr, size)
+unsafe extern "C" fn deallocate(this: NSTDAny, ptr: &mut NSTDAny, size: usize) {
+    let this = this as *mut NSTDAllocator;
+    (*this).errc |= crate::alloc::nstd_alloc_deallocate(ptr, size);
 }